requirements.txt
apache-airflow==2.0.2
apache-airflow-providers-postgres
apache-airflow-providers-redis
pandas
numpy
tensorflow
scikit-learn
mlflow
yandexcloud